+++ src/components/common/tabs/styles.ts
