+++ src/pages/challenge-record/index.tsx
@@ -2,13 +2,10 @@ import { useState } from 'react';
 
 import StampBoard from './components/stamp-board';
 import Verification from './components/verification';
-import { Tab, TabPanel, Tabs } from '@/components/common/tab';
+import { Tabs, Tab } from '@/components/common/tabs';
+import { TabPanels, TabPanel } from '@/components/common/tabs/tap-panels';
 import TopBar from '@/components/features/layout/top-bar';
 import styled from '@emotion/styled';
-
-type TabsContainerProps = {
-  position?: string;
-};
 
 const ChallengeRecord = () => {
   const [activeTab, setActiveTab] = useState<0 | 1>(0);
@@ -21,20 +18,18 @@ const ChallengeRecord = () => {
     <>
       <TopBar type='Page' title='챌린지 기록' backgroundColor='#fff' />
       <ChallengeRecordLayout>
-        <TabsContainer>
-          <Tabs selectedTab={activeTab} onChange={(value) => handleTab(value)}>
-            <Tab label='인증하기' value={0} />
-            <Tab label='기록보기' value={1} />
-          </Tabs>
-        </TabsContainer>
-        <TabPanelContainer>
+        <Tabs selectedTab={activeTab} onChange={(value) => handleTab(value)}>
+          <Tab label='인증하기' value={0} />
+          <Tab label='기록보기' value={1} />
+        </Tabs>
+        <TabPanels>
           <TabPanel value={0} selectedIndex={activeTab}>
             <Verification />
           </TabPanel>
           <TabPanel value={1} selectedIndex={activeTab}>
             <StampBoard />
           </TabPanel>
-        </TabPanelContainer>
+        </TabPanels>
       </ChallengeRecordLayout>
     </>
   );
@@ -48,23 +43,4 @@ const ChallengeRecordLayout = styled.div`
   align-items: center;
   height: 100%;
   margin: 0 1.5rem;
-`;
-
-const TabsContainer = styled.div<TabsContainerProps>`
-  display: flex;
-  align-self: center;
-  width: 100%;
-  margin: 1rem auto;
-  height: 55px;
-  border-radius: 20px;
-  background-color: var(--color-green-06);
-`;
-
-const TabPanelContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100%;
-  width: 100%;
-  text-align: center;
 `;
